Route note expressions

+++ include/sst/voicemanager/voicemanager.h
@@ -78,7 +78,7 @@ template <typename Cfg, typename Responder> struct VoiceManager
     }
 
     void routeNoteExpression(int16_t port, int16_t channel, int16_t key, int32_t noteid,
-                             int32_t expression, float value)
+                             int32_t expression, double value)
     {
         switch (voiceMode)
         {
